requirements.txt
Django
djangorestframework
django-extensions
django-mptt